LCAM-375-model moved to contribution
