+++ maat-court-data-api/src/main/java/gov/uk/courtdata/contribution/controller/ContributionsController.java
@@ -6,8 +6,8 @@ import gov.uk.courtdata.contribution.validator.CreateContributionsValidator;
 import gov.uk.courtdata.contribution.validator.UpdateContributionsValidator;
 import gov.uk.courtdata.dto.ContributionsDTO;
 import gov.uk.courtdata.dto.ErrorDTO;
-import gov.uk.courtdata.model.contribution.CreateContributions;
-import gov.uk.courtdata.model.contribution.UpdateContributions;
+import gov.uk.courtdata.contribution.model.CreateContributions;
+import gov.uk.courtdata.contribution.model.UpdateContributions;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
