+++ maat-court-data-api/src/main/java/gov/uk/courtdata/contribution/mapper/ContributionsMapper.java
@@ -2,8 +2,8 @@ package gov.uk.courtdata.contribution.mapper;
 
 import gov.uk.courtdata.dto.ContributionsDTO;
 import gov.uk.courtdata.entity.ContributionsEntity;
-import gov.uk.courtdata.model.contribution.CreateContributions;
-import gov.uk.courtdata.model.contribution.UpdateContributions;
+import gov.uk.courtdata.contribution.model.CreateContributions;
+import gov.uk.courtdata.contribution.model.UpdateContributions;
 import org.mapstruct.Mapper;
 import org.mapstruct.MappingTarget;
 import org.mapstruct.NullValuePropertyMappingStrategy;
