+++ maat-court-data-api/src/test/java/gov/uk/courtdata/contribution/service/ContributionsServiceTest.java
@@ -5,8 +5,8 @@ import gov.uk.courtdata.contribution.mapper.ContributionsMapper;
 import gov.uk.courtdata.dto.ContributionsDTO;
 import gov.uk.courtdata.entity.ContributionsEntity;
 import gov.uk.courtdata.exception.RequestedObjectNotFoundException;
-import gov.uk.courtdata.model.contribution.CreateContributions;
-import gov.uk.courtdata.model.contribution.UpdateContributions;
+import gov.uk.courtdata.contribution.model.CreateContributions;
+import gov.uk.courtdata.contribution.model.UpdateContributions;
 import gov.uk.courtdata.repository.ContributionsRepository;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
