+++ maat-court-data-api/src/test/java/gov/uk/courtdata/contribution/validator/UpdateContributionsValidatorTest.java
@@ -1,7 +1,7 @@
 package gov.uk.courtdata.contribution.validator;
 
 import gov.uk.courtdata.exception.ValidationException;
-import gov.uk.courtdata.model.contribution.UpdateContributions;
+import gov.uk.courtdata.contribution.model.UpdateContributions;
 import gov.uk.courtdata.repository.ContributionsRepository;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
